split game into class. recreated train, play, and solved

+++ py_flappybird_DQN/pipe.py
@@ -2,29 +2,32 @@ import pyglet.app
 
 class pipe():
   # x and y should be top left point of bottom quad
-  def __init__(self, x, y, width, gap, velocity, window_height, batch ):
+  def __init__(self, x, y, width, gap, velocity, window_height, batch = False ):
     self.x = x
     self.y = y
     self.width = width
     self.gap = gap
     self.velocity = velocity
     self.window_height = window_height
+    self.batch = batch
 
-    self.bottom_vertex_list = batch.add(4, pyglet.gl.GL_QUADS, None,
-      ('v2f', self.get_corners()[1]),
-      ('c3B', (0,255,0) * 4)
-    )
+    if not batch == False:
+      self.bottom_vertex_list = batch.add(4, pyglet.gl.GL_QUADS, None,
+        ('v2f', self.get_corners()[1]),
+        ('c3B', (0,255,0) * 4)
+      )
 
-    self.top_vertex_list = batch.add(4, pyglet.gl.GL_QUADS, None,
-      ('v2f', self.get_corners()[0]),
-      ('c3B', (0,255,0) * 4)
-    )
+      self.top_vertex_list = batch.add(4, pyglet.gl.GL_QUADS, None,
+        ('v2f', self.get_corners()[0]),
+        ('c3B', (0,255,0) * 4)
+      )
 
   def update(self):
     self.x -= self.velocity
 
-    self.top_vertex_list.vertices = self.get_corners()[0]
-    self.bottom_vertex_list.vertices = self.get_corners()[1]
+    if not self.batch == False:
+      self.top_vertex_list.vertices = self.get_corners()[0]
+      self.bottom_vertex_list.vertices = self.get_corners()[1]
 
   # return array with 2 values
   # val 0 is vertices for top quad
@@ -35,5 +38,6 @@ class pipe():
     return [top, bottom]
 
   def delete(self):
-    self.top_vertex_list.delete()
-    self.bottom_vertex_list.delete()
+    if not self.batch == False:
+      self.top_vertex_list.delete()
+      self.bottom_vertex_list.delete()
